+++ src/render.c
@@ -1,11 +1,14 @@
 #include "render.h"
+#include "ui.c"
 
 #define MIN_WIDTH 80
 #define MIN_HEIGHT 20
+#define MAX_WIDTH 500
+#define MAX_HEIGHT 300
 #define MAX_CHAR_AMOUNT 0x7FFFFFFF
-#define WHITEBLOCK "\033[7m \033[0m"
 
 uint16_t i;
+
 uint16_t windowSize(int8_t axis) {
 	struct winsize win;
 	ioctl(STDOUT_FILENO, TIOCGWINSZ, &win);
@@ -13,6 +16,11 @@ uint16_t windowSize(int8_t axis) {
 		puts("Minimal terminal size is 40x20.");
 		exit(1);
 	}
+	else if(win.ws_col > MAX_WIDTH || win.ws_row > MAX_HEIGHT) {
+		puts("Maximum terminal size is 500x300.");
+		exit(1);
+	}
+
 	if(axis == 'x') {
 		return win.ws_col;
 	}
@@ -25,36 +33,6 @@ void clearWindow(void) { // To provide rendering in a one frame.
 	uint16_t winHeight = windowSize('y');
 	for(i = 0; i < winHeight; i++) {
 		printf("%s", "\033[F\033[K");
-	}
-}
-
-void upperBorder(void) {
-	uint16_t winWidth = windowSize('x');
-	int8_t programName[7] = " Fiflo ";
-	uint16_t programNameCenter = (winWidth - 7) / 2;
-	for(i = 0; i < programNameCenter; i++) {
-		printf("%s", WHITEBLOCK);
-	}
-
-	printf("%s", programName);
-	if(winWidth % 2 == 0) {
-		for(i = 0; i <= programNameCenter; i++)
-		{
-			printf("%s", WHITEBLOCK);
-		}
-	}
-	else {
-		for(i = 0; i < programNameCenter; i++) {
-			printf("%s", WHITEBLOCK);
-		}
-	}
-}
-
-void lowerBorder(int32_t charCount) {
-	uint16_t winWidth = windowSize('x');
-	for(i = 0; i < winWidth; i++)
-	{
-		printf("%s", WHITEBLOCK);
 	}
 }
 
@@ -76,6 +54,6 @@ void window(int8_t pressedKey, int32_t charCount) {
 		printf("%c", '\n');
 	}
 	printf("%i\n", charCount);
-	lowerBorder(charCount);
+	lowerBorder();
 }