+++ src/render.h
@@ -8,8 +8,7 @@
 
 uint16_t windowSize(int8_t axis);
 void clearWindow(void);
-void upperBorder(void);
-void lowerBorder(int32_t charCount);
 void window(int8_t pressedKey, int32_t charCount);
+
 #endif