+++ src/ui.h
@@ -1,0 +1,12 @@
+#ifndef UI_H
+#define UI_H
+
+#include <stdio.h>
+
+#define WHITEBLOCK "\033[7m \033[0m"
+
+void upperBorder(void);
+void lowerBorder(void);
+
+#endif
+